Use IconPacks for buttons with MetroCircleButtonStyle Use AccentColorBrush for mouse over trigger
