+++ src/SimpleMusicPlayer/Resources/Icons.xaml
@@ -3,34 +3,6 @@
 
     <!-- icons taken from MahApps resources -->
 
-    <Canvas Width="48"
-            Height="48"
-            Clip="F1 M 0,0L 48,0L 48,48L 0,48L 0,0"
-            UseLayoutRounding="False"
-            x:Key="appbar_play">
-        <Path Width="12.9999"
-              Height="20"
-              Canvas.Left="19"
-              Canvas.Top="14"
-              Stretch="Fill"
-              Fill="{DynamicResource BlackBrush}"
-              Data="F1 M 19,14L 31.9999,23.7501L 31.9999,24.5001L 19,34L 19,14 Z " />
-    </Canvas>
-
-    <Canvas Width="48"
-            Height="48"
-            Clip="F1 M 0,0L 48,0L 48,48L 0,48L 0,0"
-            UseLayoutRounding="False"
-            x:Key="appbar_pause">
-        <Path Width="14"
-              Height="18"
-              Canvas.Left="17"
-              Canvas.Top="15"
-              Stretch="Fill"
-              Fill="{DynamicResource BlackBrush}"
-              Data="F1 M 17,15L 21,15L 21,33L 17,33L 17,15 Z M 27,15L 31,15L 31,33L 27,33L 27,15 Z " />
-    </Canvas>
-
     <Canvas x:Key="appbar_play_new"
             Width="48"
             Height="48">
@@ -44,100 +16,7 @@
               Fill="Black" />
     </Canvas>
 
-    <Canvas Width="48"
-            Height="48"
-            Clip="F1 M 0,0L 48,0L 48,48L 0,48L 0,0"
-            UseLayoutRounding="False"
-            x:Key="appbar_fastforward">
-        <Path Width="24"
-              Height="13"
-              Canvas.Left="12"
-              Canvas.Top="17.5"
-              Stretch="Fill"
-              Fill="{DynamicResource BlackBrush}"
-              Data="F1 M 12,17.5L 20.5,24L 12,30.5L 12,17.5 Z M 23,17.5L 31.5,24L 23,30.5L 23,17.5 Z M 33,18L 36,18L 36,30L 33,30L 33,18 Z " />
-    </Canvas>
-
-    <Canvas Width="48"
-            Height="48"
-            Clip="F1 M 0,0L 48,0L 48,48L 0,48L 0,0"
-            UseLayoutRounding="False"
-            x:Key="appbar_equalizer">
-        <Path Width="26"
-              Height="19"
-              Canvas.Left="11"
-              Canvas.Top="15"
-              Stretch="Fill"
-              Fill="{DynamicResource BlackBrush}"
-              Data="F1 M 14,15L 16,15L 16,22L 19,22L 19,25L 16,25L 16,34L 14,34L 14,25L 11,25L 11,22L 14,22L 14,15 Z M 23,15L 25,15L 25,28L 28,28L 28,31L 25,31L 25,34L 23,34L 23,31L 20,31L 20,28L 23,28L 23,15 Z M 32,15L 34,15L 34,17L 37,17L 37,20L 34,20L 34,34L 32,34L 32,20L 29,20L 29,17L 32,17L 32,15 Z " />
-    </Canvas>
-
-    <Canvas Width="48"
-            Height="48"
-            Clip="F1 M 0,0L 48,0L 48,48L 0,48L 0,0"
-            x:Key="appbar_database">
-        <Path Width="22"
-              Height="24"
-              Canvas.Left="13"
-              Canvas.Top="12"
-              Stretch="Fill"
-              Fill="{DynamicResource BlackBrush}"
-              Data="F1 M 24,12C 30.0751,12 35,13.3431 35,15C 35,15.7684 33.9409,16.4692 32.1991,17C 30.1849,17.6138 27.2578,18 24,18C 20.7422,18 17.8151,17.6138 15.8009,17C 14.0591,16.4692 13,15.7684 13,15C 13,13.3431 17.9249,12 24,12 Z M 35,33C 35,34.6569 30.0751,36 24,36C 17.9249,36 13,34.6569 13,33L 13,29C 13,30.6569 17.9249,32 24,32C 30.0751,32 35,30.6569 35,29L 35,33 Z M 35,27C 35,28.6569 30.0751,30 24,30C 17.9249,30 13,28.6569 13,27L 13,23C 13,24.6569 17.9249,26 24,26C 30.0751,26 35,24.6569 35,23L 35,27 Z M 35,21C 35,22.6569 30.0751,24 24,24C 17.9249,24 13,22.6569 13,21L 13,17C 13,18.6569 17.9249,20 24,20C 30.0751,20 35,18.6569 35,17L 35,21 Z " />
-    </Canvas>
-
-    <Canvas Width="48.0067"
-            Height="48.0067"
-            Clip="F1 M 0,0L 48.0067,0L 48.0067,48.0067L 0,48.0067L 0,0"
-            x:Key="appbar_sound_mute">
-        <Path Width="26.1606"
-              Height="17.0024"
-              Canvas.Left="11.0015"
-              Canvas.Top="16.0022"
-              Stretch="Fill"
-              Fill="{DynamicResource BlackBrush}"
-              Data="F1 M 13.0018,21.0029L 17.0024,21.0029L 23.0032,16.0022L 23.0032,33.0046L 17.0024,27.0038L 13.0018,27.0038C 11.8971,27.0038 11.0015,26.1082 11.0015,25.0035L 11.0015,23.0032C 11.0015,21.8985 11.8971,21.0029 13.0018,21.0029 Z M 25.8469,28.0394L 29.3829,24.5034L 25.8469,20.9673L 27.9685,18.8457L 31.5045,22.3818L 35.0404,18.8458L 37.1621,20.9675L 33.6261,24.5034L 37.1621,28.0394L 35.0405,30.161L 31.5045,26.625L 27.9685,30.161L 25.8469,28.0394 Z " />
-    </Canvas>
-
-    <Canvas Width="48.0067"
-            Height="48.0067"
-            Clip="F1 M 0,0L 48.0067,0L 48.0067,48.0067L 0,48.0067L 0,0"
-            x:Key="appbar_sound_3">
-        <Path Width="25.0035"
-              Height="17.0299"
-              Canvas.Left="11.0015"
-              Canvas.Top="15.9747"
-              Stretch="Fill"
-              Fill="{DynamicResource BlackBrush}"
-              Data="F1 M 13.0018,21.0029L 17.0024,21.0029L 23.0032,16.0022L 23.0032,33.0046L 17.0024,27.0038L 13.0018,27.0038C 11.8971,27.0038 11.0015,26.1082 11.0015,25.0035L 11.0015,23.0032C 11.0015,21.8985 11.8971,21.0029 13.0018,21.0029 Z M 36.005,24.0034C 36.005,27.0413 34.7736,29.7917 32.7827,31.7826L 31.3683,30.3682C 32.9972,28.7393 34.0047,26.489 34.0047,24.0034C 34.0047,21.4913 32.9757,19.2196 31.3161,17.5868L 32.5252,15.9747C 34.6669,17.982 36.005,20.8364 36.005,24.0034 Z M 32.0045,24.0034C 32.0045,26.1913 31.0008,28.1449 29.4287,29.4287L 28.0041,28.0041C 29.2186,27.0918 30.0042,25.6393 30.0042,24.0034C 30.0042,22.3674 29.2186,20.9149 28.0041,20.0026L 29.4287,18.578C 31.0008,19.8618 32.0045,21.8154 32.0045,24.0034 Z M 28.0039,24.0034C 28.0039,25.3369 27.134,26.4672 25.9306,26.8578L 25.0035,25.0035C 25.5558,25.0035 26.0036,24.5557 26.0036,24.0034C 26.0036,23.451 25.5558,23.0032 25.0035,23.0032L 25.9306,21.1489C 27.134,21.5395 28.0039,22.6698 28.0039,24.0034 Z " />
-    </Canvas>
-
-    <Canvas Width="48"
-            Height="48"
-            Clip="F1 M 0,0L 48,0L 48,48L 0,48L 0,0"
-            x:Key="appbar_add">
-        <Path Width="24"
-              Height="24"
-              Canvas.Left="12"
-              Canvas.Top="12"
-              Stretch="Fill"
-              Fill="{DynamicResource BlackBrush}"
-              Data="F1 M 22,12L 26,12L 26,22L 36,22L 36,26L 26,26L 26,36L 22,36L 22,26L 12,26L 12,22L 22,22L 22,12 Z " />
-    </Canvas>
-
     <!-- other icons -->
-
-    <Canvas x:Key="appbar_control_stop"
-            Width="48"
-            Height="48"
-            Clip="F1 M 0,0L 48,0L 48,48L 0,48L 0,0"
-            UseLayoutRounding="False">
-        <Rectangle Width="18"
-                   Height="18"
-                   Canvas.Left="15"
-                   Canvas.Top="15"
-                   Stretch="Fill"
-                   Fill="{DynamicResource BlackBrush}" />
-    </Canvas>
 
     <VisualBrush x:Key="CDBrush"
                  Stretch="Fill">
