+++ src/SimpleMusicPlayer/ValueConverters/BooleanConverter.cs
@@ -1,0 +1,44 @@
+﻿using System;
+using System.Globalization;
+using System.Windows;
+using System.Windows.Data;
+using System.Windows.Media;
+
+namespace SimpleMusicPlayer.ValueConverters
+{
+    public class BooleanConverter : Freezable, IValueConverter
+    {
+        public static readonly DependencyProperty TrueValueProperty =
+            DependencyProperty.Register("TrueValue", typeof(object), typeof(BooleanConverter), new PropertyMetadata(default(object)));
+
+        public object TrueValue
+        {
+            get { return (object) this.GetValue(TrueValueProperty); }
+            set { this.SetValue(TrueValueProperty, value); }
+        }
+
+        public static readonly DependencyProperty FalseValueProperty =
+            DependencyProperty.Register("FalseValue", typeof(object), typeof(BooleanConverter), new PropertyMetadata(default(object)));
+
+        public object FalseValue
+        {
+            get { return (object) this.GetValue(FalseValueProperty); }
+            set { this.SetValue(FalseValueProperty, value); }
+        }
+
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return (value as bool?).GetValueOrDefault() ? this.TrueValue : this.FalseValue;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return DependencyProperty.UnsetValue;
+        }
+
+        protected override Freezable CreateInstanceCore()
+        {
+            return new BooleanConverter();
+        }
+    }
+}
