+++ src/SimpleMusicPlayer/Views/EqualizerView.xaml
@@ -7,6 +7,7 @@
                                xmlns:simpleChildWindow="clr-namespace:MahApps.Metro.SimpleChildWindow;assembly=MahApps.Metro.SimpleChildWindow"
                                xmlns:player="clr-namespace:SimpleMusicPlayer.Core.Player"
                                xmlns:viewModels="clr-namespace:SimpleMusicPlayer.ViewModels"
+                               xmlns:iconPacks="http://metro.mahapps.com/winfx/xaml/iconpacks"
                                mc:Ignorable="d"
                                d:DesignHeight="300"
                                d:DesignWidth="300"
@@ -20,14 +21,7 @@
     </simpleChildWindow:ChildWindow.OverlayBrush>
 
     <simpleChildWindow:ChildWindow.Icon>
-        <Rectangle Width="22"
-                   Height="22"
-                   Fill="{DynamicResource IdealForegroundColorBrush}">
-            <Rectangle.OpacityMask>
-                <VisualBrush Stretch="Uniform"
-                             Visual="{DynamicResource appbar_equalizer}" />
-            </Rectangle.OpacityMask>
-        </Rectangle>
+        <iconPacks:PackIconEntypo Kind="SoundMix" Width="20" Height="20" HorizontalAlignment="Center" VerticalAlignment="Center" />
     </simpleChildWindow:ChildWindow.Icon>
 
     <Grid Margin="4,0,4,4">
@@ -92,3 +86,4 @@
     </Grid>
 
 </simpleChildWindow:ChildWindow>
+
